+++ cli/src/templates/project/billing-base/bootstrapper.ts
@@ -1,1 +1,0 @@
-../../../../../blueprint/billing-base/bootstrapper.ts