+++ cli/src/templates/project/worker/sdk.types.ts
@@ -1,0 +1,1 @@
+../service/sdk.types.ts