+++ blueprint/interfaces/billing/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/interfaces-billing
 
+## 0.4.0
+
+### Minor Changes
+
+- change implementation structure, and update internal dependencies, SdkClient breaking changes
+
 ## 0.3.4
 
 ### Patch Changes